refactor -> move enterRoom as handle,getConfirmationAlert to EnterRoom
